Better tree-shaking for adblocker library
