+++ web-test-runner.config.mjs
@@ -9,9 +9,6 @@ export default {
             ts: true,
             tsconfig: fileURLToPath(new URL('./tsconfig.json', import.meta.url)),
             target: 'es2021',
-            define: {
-                BUNDLE_FILTERLIST: 'false',
-            },
             json: true,
         }),
     ],
